+++ src/main/java/io/jenkins/plugins/analysis/core/util/AffectedFilesResolver.java
@@ -1,10 +1,8 @@
 package io.jenkins.plugins.analysis.core.util;
 
-import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
 import java.nio.file.Files;
-import java.nio.file.InvalidPathException;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.Set;
@@ -15,6 +13,7 @@ import edu.hm.hafner.analysis.Report;
 
 import hudson.FilePath;
 import hudson.model.Run;
+import hudson.remoting.VirtualChannel;
 
 /**
  * Copies all affected files that are referenced in at least one of the issues to Jenkins build folder. These files can
@@ -83,39 +82,46 @@ public class AffectedFilesResolver {
      *         the issues
      * @param affectedFilesFolder
      *         directory to store the copied files in
-     * @param workspace
-     *         local directory of the workspace, all source files must be part of this directory
+     * @param agentWorkspace
+     *         directory of the workspace in the agent, all source files must be part of this directory
      *
      * @throws InterruptedException
      *         if the user cancels the processing
      */
-    public void copyFilesWithAnnotationsToBuildFolder(final Report report,
-            final FilePath affectedFilesFolder, final File workspace)
-            throws InterruptedException {
+    public void copyAffectedFilesToBuildFolder(final Report report,
+            final FilePath affectedFilesFolder, final FilePath agentWorkspace) throws InterruptedException {
         int copied = 0;
         int notFound = 0;
         int notInWorkspace = 0;
 
-        FilteredLog log = new FilteredLog(report, 
+        VirtualChannel channel = agentWorkspace.getChannel();
+        FilteredLog log = new FilteredLog(report,
                 "Can't copy some affected workspace files to Jenkins build folder:");
         Set<String> files = report.getFiles();
         files.remove("-");
         for (String file : files) {
-            if (exists(file)) {
-                if (isInWorkspace(file, workspace)) {
-                    try {
-                        copy(affectedFilesFolder, file);
-                        copied++;
+            FilePath remoteFile = new FilePath(channel, file);
+            try {
+                if (remoteFile.exists()) {
+                    if (isInWorkspace(remoteFile, agentWorkspace)) {
+                        try {
+                            copy(affectedFilesFolder, file);
+                            copied++;
+                        }
+                        catch (IOException exception) {
+                            log.logError("- '%s', IO exception has been thrown: %s", file, exception);
+                        }
                     }
-                    catch (IOException exception) {
-                        log.logError("- '%s', IO exception has been thrown: %s", file, exception);
+                    else {
+                        notInWorkspace++;
                     }
                 }
                 else {
-                    notInWorkspace++;
+                    notFound++;
                 }
             }
-            else {
+            catch (IOException exception) {
+                log.logError("- '%s', IO exception has been thrown: %s", file, exception);
                 notFound++;
             }
         }
@@ -142,25 +148,11 @@ public class AffectedFilesResolver {
      *
      * @return {@code true} if the file is in the workspace, {@code false} otherwise
      */
-    private boolean isInWorkspace(final String fileName, final File workspace) {
-        try {
-            Path workspaceDirectory = workspace.toPath().toRealPath().normalize();
-            Path sourceFile = Paths.get(fileName).toRealPath();
+    private boolean isInWorkspace(final FilePath fileName, final FilePath workspace) {
+        String workspaceDirectory = workspace.getRemote();
+        String sourceFile = fileName.getRemote();
 
-            return sourceFile.startsWith(workspaceDirectory);
-        }
-        catch (IOException e) {
-            return false;
-        }
-    }
-
-    private boolean exists(final String file) {
-        try {
-            return Files.exists(Paths.get(file));
-        }
-        catch (InvalidPathException ignored) {
-            return false;
-        }
+        return sourceFile.startsWith(workspaceDirectory);
     }
 
     /**