+++ src/test/java/io/jenkins/plugins/analysis/warnings/DumbSlaveITest.java
@@ -1,25 +1,21 @@
 package io.jenkins.plugins.analysis.warnings;
 
-import java.io.ByteArrayInputStream;
-import java.io.IOException;
-import java.nio.charset.StandardCharsets;
-
 import org.junit.Test;
+import org.jvnet.hudson.test.Issue;
 
 import org.jenkinsci.plugins.workflow.cps.CpsFlowDefinition;
 import org.jenkinsci.plugins.workflow.job.WorkflowJob;
-import hudson.FilePath;
-import hudson.model.Label;
-import hudson.model.TopLevelItem;
-import hudson.security.FullControlOnceLoggedInAuthorizationStrategy;
-import hudson.security.HudsonPrivateSecurityRealm;
-import hudson.slaves.DumbSlave;
-import jenkins.security.s2m.AdminWhitelistRule;
+import hudson.model.Slave;
 
 import io.jenkins.plugins.analysis.core.model.AnalysisResult;
 import io.jenkins.plugins.analysis.core.testutil.IntegrationTestWithJenkinsPerTest;
+import io.jenkins.plugins.analysis.warnings.recorder.pageobj.DetailsTab;
+import io.jenkins.plugins.analysis.warnings.recorder.pageobj.DetailsTab.TabType;
+import io.jenkins.plugins.analysis.warnings.recorder.pageobj.IssueRow;
+import io.jenkins.plugins.analysis.warnings.recorder.pageobj.IssuesTable;
+import io.jenkins.plugins.analysis.warnings.recorder.pageobj.SourceCodeView;
 
-import static io.jenkins.plugins.analysis.core.model.AnalysisResultAssert.*;
+import static io.jenkins.plugins.analysis.core.assertions.Assertions.*;
 
 /**
  * Integration tests for builds running on a 'DumbSlave' worker node.
@@ -28,52 +24,41 @@ import static io.jenkins.plugins.analysis.core.model.AnalysisResultAssert.*;
  * @author Andreas Pabst
  */
 public class DumbSlaveITest extends IntegrationTestWithJenkinsPerTest {
+    private static final String JAVA_CONTENT = "public class Test {}";
+    private static final String C_SHARP_CONTENT = "hallo welt";
 
     /**
-     * Integration test to verify the base behaviour of building on a slave.
-     *
-     * @throws Exception
-     *         if the slave can not be created
+     * Verifies that affected source files are copied to Jenkins build folder, even if the master - agent security
+     * is active, see JENKINS-56007 for details.
      */
-    @Test
-    public void shouldBuildOnSlave() throws Exception {
-
-        DumbSlave slave = getJenkins().createOnlineSlave(Label.get("slave"));
-
-        HudsonPrivateSecurityRealm securityRealm = new HudsonPrivateSecurityRealm(false, false, null);
-        securityRealm.createAccount("user", "password");
-
-        getJenkins().jenkins.setSecurityRealm(securityRealm);
-        getJenkins().jenkins.setAuthorizationStrategy(new FullControlOnceLoggedInAuthorizationStrategy());
-
-        getJenkins().jenkins.getInjector().getInstance(AdminWhitelistRule.class).setMasterKillSwitch(false);
-        getJenkins().jenkins.save();
+    @Test @Issue("JENKINS-56007")
+    public void shouldCopySourcesIfMasterAgentSecurityIsActive() {
+        Slave agent = createAgentWithEnabledSecurity("agent");
 
         WorkflowJob project = createPipeline();
-        createFileInSlaveWorkspace(slave, project, "Test.java", "public class Test {}");
-        createFileInSlaveWorkspace(slave, project, "MediaPortal.cs", "hallo welt");
 
-        project.setDefinition(new CpsFlowDefinition("node('slave') {\n"
+        createFileInAgentWorkspace(agent, project, "Test.java", JAVA_CONTENT);
+        createFileInAgentWorkspace(agent, project, "MediaPortal.cs", C_SHARP_CONTENT);
+
+        project.setDefinition(new CpsFlowDefinition("node('agent') {\n"
                 + "    echo '[javac] Test.java:39: warning: Test Warning'\n"
                 + "    echo 'MediaPortal.cs(3001,5): warning CS0162: Hier kommt der Warnings Text'\n"
-                + "    recordIssues tools: [msBuild(), java()]\n"
+                + "    recordIssues tools: [msBuild(), java()], aggregatingResults: true\n"
                 + "}", true));
 
         AnalysisResult result = scheduleSuccessfulBuild(project);
-
         assertThat(result).hasNoErrorMessages();
-    }
+        assertThat(result).hasTotalSize(2);
 
-    private void createFileInSlaveWorkspace(final DumbSlave slave, final TopLevelItem job, final String fileName,
-            final String content) {
-        try {
-            FilePath workspace = slave.getWorkspaceFor(job);
+        DetailsTab details = new DetailsTab(getWebPage(JavaScriptSupport.JS_ENABLED, result));
 
-            FilePath child = workspace.child(fileName);
-            child.copyFrom(new ByteArrayInputStream(content.getBytes(StandardCharsets.UTF_8)));
-        }
-        catch (IOException | InterruptedException | NullPointerException e) {
-            throw new AssertionError(e);
-        }
+        IssuesTable issues = details.select(TabType.ISSUES);
+        assertThat(issues.getRows()).hasSize(2);
+
+        SourceCodeView actualCSharpContent = issues.getRow(0).click(IssueRow.FILE);
+        assertThat(actualCSharpContent.getSourceCode()).isEqualTo(C_SHARP_CONTENT);
+
+        SourceCodeView actualJavaContent = issues.getRow(1).click(IssueRow.FILE);
+        assertThat(actualJavaContent.getSourceCode()).isEqualTo(JAVA_CONTENT);
     }
 }